py/requirements.txt
numpy==1.11.1
pandas==0.18.1
patsy==0.4.1
python-firebase==1.2
requests==1.1.0
scipy==0.18.1
statsmodels==0.8.0rc1
